testcase update
