+++ index.html
@@ -14,29 +14,46 @@
   </head>
   <body>
     <div style="padding-top: 1rem; padding-left: 2rem;">
-      <form>
+      <form id="dateForm">
         <div class="container pb-2 pl-0">
-          <div class="columns col-oneline">
-            <div class="column col-6">
-              <div class="form-group">  
-                  <input class="form-input" type="date" id="dateSpec" name="dateSpec" step="any">
+              <div class="form-group">
+                  <label class="form-label" for="dateSpec">Date to process</label>
+                  <div class="columns col-oneline">
+                  <div class="column col-6">
+                  <input class="form-input form-inline" type="date" id="dateSpec" name="dateSpec" step="any">
+                  </div>
+                  <div class="column col-1 mr-2">
+                  <button class="btn form-inline" type="submit" id="startButton">Start</button>
+                  </div>
+                  <div class="column col-4 ml-2">    
+                    <label class="form-switch form-inline">
+                      <input type="checkbox" id="writeSwitch">
+                      <i class="form-icon"></i> Write to testcase
+                    </label>
+                  </div>
+                  </div>
                   <p class="form-input-hint mb-0">Only Sundays and Wednesdays are valid</p>
               </div>
-                  
-            </div>            
-            <div class="column col-1">
-              <div class="form-group"> 
-                <button class="btn" type="submit" id="startButton">Start</button>
+        </div>
+      </form>
+      <div id="orP">
+        <p style="margin-top: 12px; margin-bottom: 12px">Or…</p>
+      </div>
+      <form id="fileForm">
+        <div class="container pt-2 pb-2 pl-0">
+          <div class="form-group">
+            <label class="form-label" for="fileSpec">Testcase file to process</label>
+            <div class="columns col-oneline">
+              <div class="column col-10">
+                <input class="form-input form-inline" type="file" id="fileSpec" name="fileSpec">
               </div>
-            </div>          
-            <div class="column col-4">
-              <div class="form-group  float-right">
-                <label class="form-switch">
-                  <input type="checkbox" id="writeSwitch">
-                  <i class="form-icon"></i> Write to testcase
-                </label>
+              <!--           
+              <div class="column col-1">
+                <button class="btn form-inline" type="submit" id="startButton" disabled>Start</button>
               </div>
+              -->
             </div>
+          </div>
         </div>
       </form>
       <!-- <div id="msgs" class="bg-secondary" style="overflow: hidden"></div> -->
